docker-server/mysite/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.7
django-extensions==3.1.1
djangorestframework==3.12.2
Pillow==8.1.0
uwsgi
